refactor(release): update imports to use adapters/html
Update parser imports in remove-chart, update-qa, and update-unrc to use new adapters/html.ts path.
